separate logic and presentation from search or create, move setfocus method to logic
