+++ src/modules/elements/index.js
@@ -1,12 +1,12 @@
-export { default as Input } from './input/element.js';
-export { default as TermsList } from './terms-list/element.js';
-export { default as OrganizerForm } from './organizer-form/element.js';
-export { default as VenueForm, toFields, toVenue } from './venue-form/element.js';
-export { default as DatePicker } from './date-picker/element.js';
-export { default as TimePicker } from './time-picker/element.js';
-export { default as MetaGroup } from './meta-group/element.js';
-export { default as SearchPosts } from './search-posts/element.js';
-export { default as GoogleMap } from './google-map/element.js';
+export { default as Input } from './input/element';
+export { default as TermsList } from './terms-list/element';
+export { default as OrganizerForm } from './organizer-form/element';
+export { default as VenueForm, toFields, toVenue } from './venue-form/element';
+export { default as DatePicker } from './date-picker/element';
+export { default as TimePicker } from './time-picker/element';
+export { default as MetaGroup } from './meta-group/element';
+export { default as SearchPosts } from './search-posts/element';
+export { default as GoogleMap } from './google-map/element';
 export { default as Dashboard } from './dashboard/element';
 export { default as CheckBox } from './checkbox/element';
 export { default as Month } from './month/element';
